task1 done
took another path on this development. the simplicity of having defined element for size and array messed up my idea of code. had to do some big work on it. decided to create separated classes for each task and keep main for the output. firstly i was trying to do everything at the same page